Polkadot v0.9.32 (#356)
* polkadot-v0.9.32

* fix actions

* fix tests

* fix unit test not working && tmp remove some runtime-benchmarks test

* polkadot-v0.9.32

* Update test.yml

merge master

* fix 0.9.32 upgrade problem of election pallet && fix warning

+++ cli/tests/common.rs
@@ -1,6 +1,6 @@
 // This file is part of Substrate.
 
-// Copyright (C) 2020-2021 Parity Technologies (UK) Ltd.
+// Copyright (C) 2020-2022 Parity Technologies (UK) Ltd.
 // SPDX-License-Identifier: GPL-3.0-or-later WITH Classpath-exception-2.0
 
 // This program is free software: you can redistribute it and/or modify
@@ -19,88 +19,67 @@
 #![cfg(unix)]
 
 use assert_cmd::cargo::cargo_bin;
-use nix::sys::signal::{kill, Signal::SIGINT};
-use nix::unistd::Pid;
-use node_primitives::Block;
-use remote_externalities::rpc_api::RpcService;
-use std::{convert::TryInto, process::Command};
+use nix::{
+    sys::signal::{kill, Signal::SIGINT},
+    unistd::Pid,
+};
+use node_primitives::{Hash, Header};
 use std::{
+    io::{BufRead, BufReader, Read},
     ops::{Deref, DerefMut},
     path::Path,
-    process::{Child, ExitStatus},
-    thread,
+    process::{self, Child, Command, ExitStatus},
     time::Duration,
 };
 use tokio::time::timeout;
-static LOCALHOST_WS: &str = "ws://127.0.0.1:9944/";
 
 /// Wait for the given `child` the given number of `secs`.
 ///
 /// Returns the `Some(exit status)` or `None` if the process did not finish in the given time.
-pub fn wait_for(child: &mut Child, secs: usize) -> Option<ExitStatus> {
-    for i in 0..secs {
-        match child.try_wait().unwrap() {
-            Some(status) => {
-                if i > 5 {
-                    eprintln!("Child process took {} seconds to exit gracefully", i);
-                }
-                return Some(status);
+pub fn wait_for(child: &mut Child, secs: u64) -> Result<ExitStatus, ()> {
+    let result = wait_timeout::ChildExt::wait_timeout(child, Duration::from_secs(5.min(secs)))
+        .map_err(|_| ())?;
+    if let Some(exit_status) = result {
+        Ok(exit_status)
+    } else {
+        if secs > 5 {
+            eprintln!("Child process taking over 5 seconds to exit gracefully");
+            let result = wait_timeout::ChildExt::wait_timeout(child, Duration::from_secs(secs - 5))
+                .map_err(|_| ())?;
+            if let Some(exit_status) = result {
+                return Ok(exit_status);
             }
-            None => thread::sleep(Duration::from_secs(1)),
         }
+        eprintln!("Took too long to exit (> {} seconds). Killing...", secs);
+        let _ = child.kill();
+        child.wait().unwrap();
+        Err(())
     }
-    eprintln!("Took too long to exit (> {} seconds). Killing...", secs);
-    let _ = child.kill();
-    child.wait().unwrap();
-
-    None
-}
-
-/// Run the node for a while (60 seconds)
-pub fn run_dev_node_for_a_while(base_path: &Path) {
-    let mut cmd = Command::new(cargo_bin("deeper-chain"));
-
-    let mut cmd = cmd
-        .args(&["--dev"])
-        .arg("-d")
-        .arg(base_path)
-        .spawn()
-        .unwrap();
-
-    // Let it produce some blocks.
-    thread::sleep(Duration::from_secs(60));
-    assert!(
-        cmd.try_wait().unwrap().is_none(),
-        "the process should still be running"
-    );
-
-    // Stop the process
-    kill(Pid::from_raw(cmd.id().try_into().unwrap()), SIGINT).unwrap();
-    assert!(wait_for(&mut cmd, 40)
-        .map(|x| x.success())
-        .unwrap_or_default());
 }
 
 /// Wait for at least n blocks to be finalized within a specified time.
 pub async fn wait_n_finalized_blocks(
     n: usize,
     timeout_secs: u64,
+    url: &str,
 ) -> Result<(), tokio::time::error::Elapsed> {
     timeout(
         Duration::from_secs(timeout_secs),
-        wait_n_finalized_blocks_from(n, LOCALHOST_WS),
+        wait_n_finalized_blocks_from(n, url),
     )
     .await
 }
 
 /// Wait for at least n blocks to be finalized from a specified node
 pub async fn wait_n_finalized_blocks_from(n: usize, url: &str) {
+    use substrate_rpc_client::{ws_client, ChainApi};
+
     let mut built_blocks = std::collections::HashSet::new();
     let mut interval = tokio::time::interval(Duration::from_secs(2));
-    let rpc_service = RpcService::new(url, false).await.unwrap();
+    let rpc = ws_client(url).await.unwrap();
 
     loop {
-        if let Ok(block) = rpc_service.get_finalized_head::<Block>().await {
+        if let Ok(block) = ChainApi::<(), Hash, Header, ()>::finalized_head(&rpc).await {
             built_blocks.insert(block);
             if built_blocks.len() > n {
                 break;
@@ -108,6 +87,50 @@ pub async fn wait_n_finalized_blocks_from(n: usize, url: &str) {
         };
         interval.tick().await;
     }
+}
+
+/// Run the node for a while (3 blocks)
+pub async fn run_node_for_a_while(base_path: &Path, args: &[&str]) {
+    let mut cmd = Command::new(cargo_bin("deeper-chain"))
+        .stdout(process::Stdio::piped())
+        .stderr(process::Stdio::piped())
+        .args(args)
+        .arg("-d")
+        .arg(base_path)
+        .spawn()
+        .unwrap();
+
+    let stderr = cmd.stderr.take().unwrap();
+
+    let mut child = KillChildOnDrop(cmd);
+
+    let (ws_url, _) = find_ws_url_from_output(stderr);
+
+    // Let it produce some blocks.
+    let _ = wait_n_finalized_blocks(3, 30, &ws_url).await;
+
+    assert!(
+        child.try_wait().unwrap().is_none(),
+        "the process should still be running"
+    );
+
+    // Stop the process
+    kill(Pid::from_raw(child.id().try_into().unwrap()), SIGINT).unwrap();
+    assert!(wait_for(&mut child, 40).map(|x| x.success()).unwrap());
+}
+
+/// Run the node asserting that it fails with an error
+pub fn run_node_assert_fail(base_path: &Path, args: &[&str]) {
+    let mut cmd = Command::new(cargo_bin("deeper-chain"));
+
+    let mut child = KillChildOnDrop(cmd.args(args).arg("-d").arg(base_path).spawn().unwrap());
+
+    // Let it produce some blocks, but it should die within 10 seconds.
+    assert_ne!(
+        wait_timeout::ChildExt::wait_timeout(&mut *child, Duration::from_secs(10)).unwrap(),
+        None,
+        "the process should not be running anymore"
+    );
 }
 
 pub struct KillChildOnDrop(pub Child);
@@ -130,4 +153,31 @@ impl DerefMut for KillChildOnDrop {
     fn deref_mut(&mut self) -> &mut Self::Target {
         &mut self.0
     }
+}
+
+/// Read the WS address from the output.
+///
+/// This is hack to get the actual binded sockaddr because
+/// substrate assigns a random port if the specified port was already binded.
+pub fn find_ws_url_from_output(read: impl Read + Send) -> (String, String) {
+    let mut data = String::new();
+
+    let ws_url = BufReader::new(read)
+        .lines()
+        .find_map(|line| {
+            let line =
+                line.expect("failed to obtain next line from stdout for WS address discovery");
+            data.push_str(&line);
+
+            // does the line contain our port (we expect this specific output from substrate).
+            let sock_addr = match line.split_once("Running JSON-RPC WS server: addr=") {
+                None => return None,
+                Some((_, after)) => after.split_once(",").unwrap().0,
+            };
+
+            Some(format!("ws://{}", sock_addr))
+        })
+        .expect("We should get a WebSocket address");
+
+    (ws_url, data)
 }
